refactor: update SignatureDescriptor import path to reduce type issues
